models/candidate.py
from extensions import db

class Candidate(db.Model):
    __tablename__ = 'candidates'
    id = db.Column(db.Integer, primary_key=True)
    name = db.Column(db.String(120), nullable=False)
    party = db.Column(db.String(120))
    votes = db.Column(db.Integer, default=0)
